variables and partial tests
