+++ policyengine_canada/tests/gov/provinces/sk/benefits/afb/sk_active_family_benefit.yaml
@@ -1,0 +1,6 @@
+- name: Income threshold
+  period: 2022
+  input: 
+    province_code: SK
+    income: 65000
+  output:
